Tests resources

+++ src/main/java/com/cesi/ressourcesrelationnelles/service/RessourceService.java
@@ -1,15 +1,15 @@
 package com.cesi.ressourcesrelationnelles.service;
 
 import com.cesi.ressourcesrelationnelles.domain.Resource;
+import com.cesi.ressourcesrelationnelles.exception.ResourceAlreadyExistException;
 import com.cesi.ressourcesrelationnelles.exception.ResourceNotFoundException;
+import com.cesi.ressourcesrelationnelles.exception.ResourceNotValidException;
 import com.cesi.ressourcesrelationnelles.repository.RessourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,11 +28,10 @@ public class RessourceService {
         for (Resource item : iterable) {
             resources.add(item);
         }
-        //List<Ressource> resources = resourceRepository.findAll();
         return resources;
     }
 
-    public List<Resource> list(Date date){
+    public List<Resource> list(Date date) throws ResourceNotFoundException {
         List<Resource> resources = new ArrayList<>();
         Iterable<Resource> iterable = resourceRepository.findAll();
         for (Resource item : iterable) {
@@ -41,6 +40,9 @@ public class RessourceService {
             System.out.println("date= " + item.getPublishDate());
         }
         resources = resources.stream().filter(resource -> resource.getPublishDate().compareTo(date) == (0)).collect(Collectors.toList());
+        if(resources.size() == 0){
+            throw new ResourceNotFoundException();
+        }
         return resources;
     }
 
@@ -52,18 +54,61 @@ public class RessourceService {
         throw new ResourceNotFoundException();
     }
 
-    public Resource createResource(Resource resource){
-        return resourceRepository.save(resource);
+    public Resource createResource(Resource resource) throws ResourceNotValidException, ResourceAlreadyExistException, ResourceNotFoundException {
+        if(resource.getId() == null || resource.getTitle() == null || resource.getPublishDate() == null || resource.getUrlContent() == null || resource.getViewNumber() == null || resource.getAuthor() == null || resource.getType() == null || resource.getCategory() == null){
+            throw new ResourceNotValidException();
+        }
+        if(this.idExists(resource.getId())){
+            throw new ResourceAlreadyExistException();
+        }
+        try {
+            return resourceRepository.save(resource);
+        }catch(JpaObjectRetrievalFailureException e){
+            throw new ResourceNotValidException();
+        }
     }
 
-    public Resource updateResource(Resource resource) throws ResourceNotFoundException {
-        if(resourceRepository.findById(resource.getId()).isPresent()){
+    private boolean idExists(Long id) {
+        Optional<Resource> res = resourceRepository.findById(id);
+        if(res.isPresent()){
+            return true;
+        }
+        return false;
+    }
+
+    public Resource updateResource(Resource resource) throws ResourceNotFoundException, ResourceNotValidException {
+        Optional<Resource> resOpt = resourceRepository.findById(resource.getId());
+        if(resOpt.isPresent()){
+            Resource old = resOpt.get();
+            if(old.getPublishDate().compareTo(resource.getPublishDate()) != 0){
+                throw new ResourceNotValidException();
+            }
             return resourceRepository.save(resource);
         }
         throw new ResourceNotFoundException();
     }
 
-    public void deleteResource(long id){
-        resourceRepository.deleteById(id);
+    public void deleteResource(long id) throws ResourceNotFoundException {
+        if(this.idExists(id)) {
+            resourceRepository.deleteById(id);
+        }
+        else {
+            throw new ResourceNotFoundException();
+        }
+    }
+
+    public List<Resource> listOrderByDateDesc() {
+        List<Resource> res = this.list();
+        res.sort(Comparator.comparing(Resource::getPublishDate));
+        return res;
+    }
+
+    public List<Resource> search(String reseachText) throws ResourceNotFoundException {
+        List<Resource> res = this.list();
+        res = res.stream().filter(resource -> resource.getTitle().contains(reseachText)).collect(Collectors.toList());
+        if(res.size() == 0){
+            throw new ResourceNotFoundException();
+        }
+        return res;
     }
 }
